added collaboration v4, but not fully functional yet, fixed some bugs

+++ public/javascripts/delete_book.js
@@ -3,10 +3,15 @@ const id = deleteButton.getAttribute("data-id")
 deleteButton.onclick = async (e) => {
 	e.preventDefault() 	// do not trigger default functionality		
 	console.log(id)
-	await fetch(`http://localhost:3000/books/delete/${id}`, {
-		method: 'DELETE'
-	})
-	window.location.href = '/users/profile';
+	if (confirm("Are you sure you want to delete this? Once deleted all pages of this book are permanently deleted as well. Continue?")) {
+		await fetch(`http://localhost:3000/books/delete/${id}`, {
+			method: 'DELETE'
+		})
+		window.location.href = '/users/profile';
+	} else {
+		window.location.href = '/users/profile';
+	}
+
 }
 
 // headers: {
