+++ public/javascripts/delete_page.js
@@ -4,10 +4,15 @@ const bookid = deleteButton.getAttribute("data-bookid")
 deleteButton.onclick = async (e) => {
 	e.preventDefault() 	// do not trigger default functionality		
 	console.log(id)
-	await fetch(`http://localhost:3000/books/page/delete/${id}`, {
-		method: 'DELETE'
-	})
-	window.location.href = `/books/${bookid}`;
+	if (confirm("Are you sure you want to delete this page?")) {
+		await fetch(`http://localhost:3000/books/page/delete/${id}`, {
+			method: 'DELETE'
+		})
+		window.location.href = `/books/${bookid}`;
+	} else {
+		window.location.href = `/books/${bookid}`;
+	}
+
 }
 
 // headers: {
